+++ script.js
@@ -1,11 +1,6 @@
-import fs from 'fs';
-import path from 'path';
 import os from 'os';
 import readline from 'readline';
-import crypto from 'crypto';
-import zlib from 'zlib';
 import {stdin as input, stdout as output, chdir} from 'process';
-import { pipeline } from 'stream';
 import {showInfo} from'./shared.js';
 import {up} from './nwd/up.js';
 import {cd} from './nwd/cd.js';
@@ -18,6 +13,8 @@ import {mv} from './basic/mv.js';
 import {rmFile} from './basic/rm.js';
 import {osFunc} from './os/os.js';
 import {hash} from './hash/hash.js'
+import { compress } from './zip/compress.js';
+import { decompress } from './zip/decompress.js';
 
 export const errorText = 'Operation failed\n';
 const readLine = readline.createInterface({input, output});
@@ -27,60 +24,6 @@ const exit = () => {
     output.write(`Thank you for using File Manager, ${userName}!`);
     output.end();
     readLine.close();
-}
-
-const compress = async (data) => {
-    const pathToFile = data.slice(0, data.indexOf(' ')); 
-    const pathToDestination = data.slice(data.indexOf(' ')+1); 
-    fs.access(pathToDestination,  (err) => {
-        if(!err) {
-            output.write('1');
-            showInfo();
-        } else {
-            const read = fs.createReadStream(pathToFile);
-            const write = fs.createWriteStream(pathToDestination);
-            const brotliCompress = zlib.createBrotliCompress();
-            fs.access(pathToFile, (err) => {
-                if(err) {
-                    output.write('2');
-                    showInfo();
-                }
-                else{
-                    pipeline(read, brotliCompress, write, err => {
-                        if(err) output.write('3');
-                        rmFile(pathToFile);
-                    });
-                }
-            })
-        }
-    })
-}
-
-const decompress = (data) => {
-    const pathToFile = data.slice(0, data.indexOf(' ')); 
-    const pathToDestination = data.slice(data.indexOf(' ')+1); 
-    fs.access(pathToDestination,  (err) => {
-        if(!err) {
-            output.write(errorText);
-            showInfo();
-        } else {
-            const read = fs.createReadStream(pathToFile);
-            const write = fs.createWriteStream(pathToDestination);
-            const brotliDecompress = zlib.createBrotliDecompress();
-            fs.access(pathToFile, (err) => {
-                if(err) {
-                    output.write(errorText);
-                    showInfo();
-                }
-                else{
-                    pipeline(read, brotliDecompress, write, err => {
-                        if(err) output.write(errorText);
-                        rmFile(pathToFile);
-                    });
-                }
-            })
-        }
-    })
 }
 
 readLine.on('line', async input => {
